deu pane no migrations.ts
