--- modulo5CERTO/lama/lama/.gitignore
@@ -1,5 +1,0 @@
-.env
-.DS_STORE
-node_modules
-build
-coverage
