--- modulo5CERTO/lama/lama/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "projeto-cookenu-backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "node ./build/src/index.js",
-    "build": "tsc",
-    "dev": "ts-node-dev ./src/index.ts",
-    "migrations": "tsc && node ./build/src/database/migrations/Migrations.js",
-    "test": "jest"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/bcryptjs": "^2.4.2",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.13",
-    "@types/jest": "^28.1.6",
-    "@types/jsonwebtoken": "^8.5.8",
-    "@types/knex": "^0.16.1",
-    "@types/node": "^18.0.6",
-    "@types/uuid": "^8.3.4",
-    "jest": "^28.1.3",
-    "ts-jest": "^28.0.8",
-    "ts-node-dev": "^2.0.0",
-    "typescript": "^4.7.4"
-  },
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.1",
-    "express": "^4.18.1",
-    "jsonwebtoken": "^8.5.1",
-    "knex": "^2.2.0",
-    "mysql": "^2.18.1",
-    "uuid": "^8.3.2"
-  }
-}
