--- modulo5CERTO/lama/lama/requests.rest
@@ -1,3 +1,0 @@
-### Endpoint de teste
-GET http://localhost:3003/ping
-
