--- modulo5CERTO/lama/lama/src/business/ShowBusiness.ts
@@ -1,13 +1,0 @@
-import { ShowDatabase } from "../database/ShowDatabase"
-import { Authenticator } from "../services/Authenticator"
-import { HashManager } from "../services/HashManager"
-import { IdGenerator } from "../services/IdGenerator"
-
-export class ShowBusiness {
-    constructor(
-        private showDatabase: ShowDatabase,
-        private idGenerator: IdGenerator,
-        private authenticator: Authenticator
-    ) {}
-
-}
