--- modulo5CERTO/lama/lama/src/business/UserBusiness.ts
@@ -1,14 +1,0 @@
-import { UserDatabase } from "../database/UserDatabase"
-import { Authenticator } from "../services/Authenticator"
-import { HashManager } from "../services/HashManager"
-import { IdGenerator } from "../services/IdGenerator"
-
-export class UserBusiness {
-    constructor(
-        private userDatabase: UserDatabase,
-        private idGenerator: IdGenerator,
-        private hashManager: HashManager,
-        private authenticator: Authenticator
-    ) {}
-
-}
