--- modulo5CERTO/lama/lama/src/controller/ShowController.ts
@@ -1,9 +1,0 @@
-import { Request, Response } from "express";
-import { ShowBusiness } from "../business/ShowBusiness";
-
-export class ShowController {
-    constructor(
-        private showBusiness: ShowBusiness
-    ) {}
-
-}
