--- modulo5CERTO/lama/lama/src/controller/UserController.ts
@@ -1,9 +1,0 @@
-import { Request, Response } from "express";
-import { UserBusiness } from "../business/UserBusiness";
-
-export class UserController {
-    constructor(
-        private userBusiness: UserBusiness
-    ) {}
-
-}
