--- modulo5CERTO/lama/lama/src/database/ShowDatabase.ts
@@ -1,7 +1,0 @@
-import { BaseDatabase } from "./BaseDatabase"
-
-export class ShowDatabase extends BaseDatabase {
-    public static TABLE_SHOWS = "Lama_Shows"
-    public static TABLE_TICKETS = "Lama_Tickets"
-
-}
