--- modulo5CERTO/lama/lama/src/database/UserDatabase.ts
@@ -1,6 +1,0 @@
-import { BaseDatabase } from "./BaseDatabase"
-
-export class UserDatabase extends BaseDatabase {
-    public static TABLE_USERS = "Lama_Users"
-
-}
