--- modulo5CERTO/lama/lama/src/errors/RequestError.ts
@@ -1,9 +1,0 @@
-import { BaseError } from "./BaseError";
-
-export class RequestError extends BaseError {
-    constructor(
-        message: string = "Requisição inválida"
-    ) {
-        super(400, message)
-    }
-}
