--- modulo5CERTO/lama/lama/src/errors/UnauthorizedError.ts
@@ -1,9 +1,0 @@
-import { BaseError } from "./BaseError";
-
-export class UnauthorizedError extends BaseError {
-    constructor(
-        message: string = "Credenciais inválidas"
-    ) {
-        super(401, message)
-    }
-}
