--- modulo5CERTO/lama/lama/src/router/userRouter.ts
@@ -1,19 +1,0 @@
-import { Router } from 'express'
-import { UserBusiness } from '../business/UserBusiness'
-import { UserController } from '../controller/UserController'
-import { UserDatabase } from '../database/UserDatabase'
-import { Authenticator } from '../services/Authenticator'
-import { HashManager } from '../services/HashManager'
-import { IdGenerator } from '../services/IdGenerator'
-
-export const userRouter = Router()
-
-const userController = new UserController(
-    new UserBusiness(
-        new UserDatabase(),
-        new IdGenerator(),
-        new HashManager(),
-        new Authenticator()
-    )
-)
-
