--- modulo5CERTO/lama/lama/tests/PingBusinessTests/ping.test.ts
@@ -1,12 +1,0 @@
-import { PingBusiness } from "../../src/business/PingBusiness"
-
-describe("Testing PingBusiness", () => {
-    const pingBusiness = new PingBusiness(
-        // não possui dependências
-    )
-    
-    test("deve retornar 'Pong!' em caso de sucesso", async () => {
-        const response = await pingBusiness.ping()
-        expect(response.message).toBe("Pong!")
-    })
-})
