--- modulo5CERTO/lama/lama/tests/mocks/ShowDatabaseMock.ts
@@ -1,7 +1,0 @@
-import { BaseDatabase } from "../../src/database/BaseDatabase"
-
-export class ShowDatabaseMock extends BaseDatabase {
-    public static TABLE_POSTS = "Labook_Posts"
-    public static TABLE_LIKES = "Labook_Likes"
-
-}
