--- modulo5CERTO/lama/lama/tests/mocks/UserDatabaseMock.ts
@@ -1,6 +1,0 @@
-import { BaseDatabase } from "../../src/database/BaseDatabase"
-
-export class UserDatabaseMock extends BaseDatabase {
-    public static TABLE_USERS = "Labook_Users"
-
-}
