--- modulo5CERTO/lama/lama/tests/mocks/services/IdGeneratorMock.ts
@@ -1,5 +1,0 @@
-export class IdGeneratorMock {
-    public generate = (): string => {
-        return "id-mock"
-    }
-}
