--- modulo5CERTO/lama/lama/tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es2021",
-        "module": "commonjs",
-        "sourceMap": true,
-        "outDir": "./build",
-        "rootDir": "./",
-        "removeComments": true,
-        "noImplicitAny": true,
-        "esModuleInterop": true,
-        "strict": true
-    },
-    "exclude": [
-        "./tests/*"
-    ]
-}
