+++ modulo5CERTO/lama/requests.rest
@@ -1,0 +1,2 @@
+### Ping
+GET http://localhost:3003/ping
