+++ modulo5CERTO/lama/src/business/PingBusiness.ts
