+++ modulo5CERTO/lama/src/database/BaseDatabase.ts
@@ -1,0 +1,18 @@
+import knex from "knex"
+import dotenv from "dotenv"
+
+dotenv.config()
+
+export abstract class BaseDatabase {
+    protected static connection = knex({
+        client: "mysql",
+        connection: {
+            host: process.env.DB_HOST,
+            port: 3306,
+            user: process.env.DB_USER,
+            password: process.env.DB_PASSWORD,
+            database: process.env.DB_DATABASE,
+            multipleStatements: true
+        },
+    })
+}
