+++ modulo5CERTO/lama/src/database/UserDatabase.ts
@@ -1,0 +1,35 @@
+import { IUserDB, User } from "../models/User"
+import { BaseDatabase } from "./BaseDatabase"
+
+export class UserDatabase extends BaseDatabase {
+    public static TABLE_USERS = "Template_Users"
+
+    public toUserDBModel = (user: User): IUserDB => {
+        const userDB: IUserDB = {
+            id: user.getId(),
+            name: user.getName(),
+            email: user.getEmail(),
+            password: user.getPassword(),
+            role: user.getRole()
+        }
+
+        return userDB
+    }
+
+    public findByEmail = async (email: string): Promise<IUserDB | undefined> => {
+        const result: IUserDB[] = await BaseDatabase
+            .connection(UserDatabase.TABLE_USERS)
+            .select()
+            .where({ email })
+
+        return result[0]
+    }
+
+    public createUser = async (user: User): Promise<void> => {
+        const userDB = this.toUserDBModel(user)
+
+        await BaseDatabase
+            .connection(UserDatabase.TABLE_USERS)
+            .insert(userDB)
+    }
+}
