+++ modulo5CERTO/lama/src/errors/AuthenticationError.ts
