+++ modulo5CERTO/lama/src/errors/AuthorizationError.ts
