+++ modulo5CERTO/lama/src/errors/BaseError.ts
