+++ modulo5CERTO/lama/src/errors/ConflictError.ts
