+++ modulo5CERTO/lama/src/errors/NotFoundError.ts
