+++ modulo5CERTO/lama/src/errors/ParamsError.ts
