+++ modulo5CERTO/lama/src/router/pingRouter.ts
